+++ lib/features/logger/ui/logs_list/screen/desktop_logs_list_screen.dart
@@ -24,6 +24,14 @@ class _DesktopLogsListScreenState extends State<DesktopLogsListScreen> {
   void initState() {
     super.initState();
     widget.notifier.addListener(_onNotifierChanged);
+    widget.notifier.onShareAllLogs = (sharableFile) {
+      if (Infospect.instance.onShareAllLogs != null) {
+        Infospect.instance.onShareAllLogs!(sharableFile.path);
+      } else {
+        final XFile file = XFile(sharableFile.path);
+        SharePlus.instance.share(ShareParams(files: [file]));
+      }
+    };
   }
 
   void _onNotifierChanged() {
@@ -33,24 +41,11 @@ class _DesktopLogsListScreenState extends State<DesktopLogsListScreen> {
   @override
   void dispose() {
     widget.notifier.removeListener(_onNotifierChanged);
-    // Don't dispose notifier here - it's managed by navigation_helper
     super.dispose();
   }
 
   @override
   Widget build(BuildContext context) {
-    if (widget.notifier.sharableFile != null) {
-      WidgetsBinding.instance.addPostFrameCallback((_) {
-        if (Infospect.instance.onShareAllLogs != null) {
-          Infospect
-              .instance.onShareAllLogs!(widget.notifier.sharableFile!.path);
-        } else {
-          final XFile file = XFile(widget.notifier.sharableFile!.path);
-          SharePlus.instance.share(ShareParams(files: [file]));
-        }
-      });
-    }
-
     return Scaffold(
       appBar: LogsListAppBar.desktop(
         hasBottom: widget.notifier.filters.isNotEmpty,