+++ lib/features/network/ui/list/screen/desktop_networks_list_screen.dart
@@ -34,6 +34,14 @@ class _DesktopNetworksListScreenState extends State<DesktopNetworksListScreen> {
   void initState() {
     super.initState();
     widget.notifier.addListener(_onNotifierChanged);
+    widget.notifier.onShareAllNetworkCalls = (sharableFile) {
+      if (Infospect.instance.onShareAllNetworkCalls != null) {
+        Infospect.instance.onShareAllNetworkCalls!(sharableFile.path);
+      } else {
+        final XFile file = XFile(sharableFile.path);
+        SharePlus.instance.share(ShareParams(files: [file]));
+      }
+    };
   }
 
   void _onNotifierChanged() {
@@ -53,24 +61,11 @@ class _DesktopNetworksListScreenState extends State<DesktopNetworksListScreen> {
   @override
   void dispose() {
     widget.notifier.removeListener(_onNotifierChanged);
-    // Don't dispose notifier here - it's managed by navigation_helper
     super.dispose();
   }
 
   @override
   Widget build(BuildContext context) {
-    if (widget.notifier.sharableFile != null) {
-      WidgetsBinding.instance.addPostFrameCallback((_) {
-        if (Infospect.instance.onShareAllNetworkCalls != null) {
-          Infospect.instance
-              .onShareAllNetworkCalls!(widget.notifier.sharableFile!.path);
-        } else {
-          final XFile file = XFile(widget.notifier.sharableFile!.path);
-          Share.shareXFiles([file]);
-        }
-      });
-    }
-
     return Scaffold(
       appBar: NetworkCallAppBar.desktop(
         hasBottom: widget.notifier.filters.isNotEmpty,