+++ lib/utils/data_transfer.dart
@@ -1,0 +1,6 @@
+class InfospectDataTransfer {
+  static String channelName = 'infospect_channel';
+  static String windowName = 'Infospect';
+  static String onSend = 'onSend';
+  static String broadcast = 'broadcast';
+}